Move DayComponents to own file for expansion
